# docs/swagger.yaml
openapi: 3.0.4
info:
  title: E-commerce API
  description: |
    Complete e-commerce API with user authentication, product management, 
    shopping cart, orders, payments, and admin functionality.
    
    ## Authentication
    - JWT Bearer token authentication
    - Email verification required for full access
    - Google OAuth support
    
    ## Roles
    - **user**: Regular customer permissions
    - **admin**: Full administrative access
    
    ## Rate Limiting
    - 1000 requests per hour per IP
    - 100 requests per minute for authenticated users
  version: 1.0.0
  termsOfService: https://example.com/terms
  contact:
    name: Abdulrahman Hegazy
    email: a.0hegazy@proton.me
...
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://ecommerce-express-boilerplate.onrender.com/api/v1
    description: Production server

# Global security - applies to all endpoints
security:
  - bearerAuth: []
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve paginated products with filtering, sorting, and search capabilities
      operationId: getAllProducts
      parameters:
        - $ref: '#/components/parameters/PaginationPage'
        - $ref: '#/components/parameters/PaginationLimit'
        - in: query
          name: category
          schema:
            type: string
          description: Filter by product category
        - in: query
          name: minPrice
          schema:
            type: number
            format: float
            minimum: 0
          description: Minimum price filter
        - in: query
          name: maxPrice
          schema:
            type: number
            format: float
            minimum: 0
          description: Maximum price filter
        - in: query
          name: search
          schema:
            type: string
          description: Search term for product name and description
        - in: query
          name: featured
          schema:
            type: boolean
          description: Filter featured products only
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, rating, name, createdAt]
          description: Sort products by criteria
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Products retrieved
                  value:
                    status: success
                    data:
                      products:
                        - id: "64c9b8d3a7e9a2d8f0c1b236"
                          name: "Wireless Headphones"
                          description: "High-quality wireless headphones with noise cancellation"
                          category: "electronics"
                          images: 
                            - "https://example.com/headphones1.jpg"
                            - "https://example.com/headphones2.jpg"
                          price: 29.99
                          originalPrice: 39.99
                          stock: 50
                          brand: "AudioTech"
                          features:
                            - "Noise Cancellation"
                            - "Bluetooth 5.0"
                            - "20h Battery"
                          ratings:
                            average: 4.5
                            count: 120
                          tags: ["wireless", "audio", "headphones"]
                          isActive: true
                          isFeatured: true
                          createdAt: "2025-10-15T10:00:00.000Z"
                          updatedAt: "2025-10-15T12:00:00.000Z"
                      pagination:
                        total: 150
                        currentPage: 1
                        totalPages: 13
                        pageSize: 12
        '500':
          $ref: '#/components/responses/ServerError'
  /products/featured:
    get:
      tags:
        - Products
      summary: Get featured products
      description: Retrieve featured products sorted by rating
      operationId: getFeaturedProducts
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 8
            minimum: 1
            maximum: 20
          description: Number of featured products to return
      responses:
        '200':
          description: Featured products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Featured products
                  value:
                    status: success
                    data:
                      products:
                        - id: "64c9b8d3a7e9a2d8f0c1b236"
                          name: "Wireless Headphones"
                          description: "High-quality wireless headphones"
                          category: "electronics"
                          images: ["https://example.com/headphones.jpg"]
                          price: 29.99
                          originalPrice: 39.99
                          stock: 50
                          brand: "AudioTech"
                          ratings:
                            average: 4.5
                            count: 120
                          isActive: true
                          isFeatured: true
                        - id: "64c9b8d3a7e9a2d8f0c1b237"
                          name: "Smart Watch"
                          description: "Feature-rich smartwatch"
                          category: "electronics"
                          images: ["https://example.com/watch.jpg"]
                          price: 199.99
                          originalPrice: 249.99
                          stock: 25
                          brand: "TechWear"
                          ratings:
                            average: 4.2
                            count: 89
                          isActive: true
                          isFeatured: true
        '500':
          $ref: '#/components/responses/ServerError'
  /products/categories:
    get:
      tags:
        - Products
      summary: Get product categories
      description: Retrieve all available product categories with product counts
      operationId: getProductCategories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Categories list
                  value:
                    status: success
                    data:
                      categories:
                        - category: "electronics"
                          count: 45
                        - category: "clothing"
                          count: 78
                        - category: "home"
                          count: 32
                        - category: "sports"
                          count: 23
        '500':
          $ref: '#/components/responses/ServerError'
  /products/search:
    get:
      tags:
        - Products
      summary: Search products
      description: Search products with advanced filtering and sorting
      operationId: searchProducts
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Search query text
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category
        - in: query
          name: minPrice
          schema:
            type: number
            format: float
          description: Minimum price
        - in: query
          name: maxPrice
          schema:
            type: number
            format: float
          description: Maximum price
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, rating, name, createdAt]
          description: Sort results
        - $ref: '#/components/parameters/PaginationPage'
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Search results
                  value:
                    status: success
                    data:
                      products:
                        - id: "64c9b8d3a7e9a2d8f0c1b236"
                          name: "Wireless Headphones"
                          description: "High-quality wireless headphones"
                          category: "electronics"
                          images: ["https://example.com/headphones.jpg"]
                          price: 29.99
                          stock: 50
                          brand: "AudioTech"
                          ratings:
                            average: 4.5
                            count: 120
                          isActive: true
                          isFeatured: true
                      pagination:
                        total: 15
                        currentPage: 1
                        totalPages: 2
                        pageSize: 12
        '500':
          $ref: '#/components/responses/ServerError'
  /products/{productId}:
    get:
      tags:
        - Products
      summary: Get single product
      description: Retrieve detailed information about a specific product
      operationId: getSingleProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Product details
                  value:
                    status: success
                    data:
                      product:
                        id: "64c9b8d3a7e9a2d8f0c1b236"
                        name: "Wireless Headphones"
                        description: "High-quality wireless headphones with active noise cancellation technology. Perfect for travel and office use."
                        category: "electronics"
                        images:
                          - "https://example.com/headphones1.jpg"
                          - "https://example.com/headphones2.jpg"
                          - "https://example.com/headphones3.jpg"
                        price: 29.99
                        originalPrice: 39.99
                        stock: 50
                        brand: "AudioTech"
                        features:
                          - "Active Noise Cancellation"
                          - "Bluetooth 5.0"
                          - "20-hour battery life"
                          - "Quick charge (15 min = 3 hours)"
                          - "Built-in microphone"
                        ratings:
                          average: 4.5
                          count: 120
                        tags: ["wireless", "audio", "headphones", "bluetooth"]
                        isActive: true
                        isFeatured: true
                        createdAt: "2025-10-15T10:00:00.000Z"
                        updatedAt: "2025-10-15T12:00:00.000Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. For admin endpoints, user must have 'admin' role.
  parameters:
    PaginationPage:
      in: query
      name: page
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    PaginationLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: Number of items per page
    ProductId:
      in: path
      name: productId
      required: true
      schema:
        type: string
        format: objectid
      description: Product ID
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Resource not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data: "Internal Server Error"
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, fail, error]
          example: success
        data:
          oneOf:
            - type: object
              description: Contains response data in success and fail cases
              properties:
                message:
                  type: string
                  example: "Successfully completed action"
                accessToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5..."
                user:
                  $ref: '#/components/schemas/User'
            - type: string
              description: Contains error message or additional info in error cases
              example: "Internal server error"
            - type: object
              nullable: true
              description: If no data is available, null can be provided
              example: null
      required:
        - status
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        data:
          type: string
          example: "Internal Server Error"
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "64c9b8d3a7e9a2d8f0c1b234"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        avatar:
          type: string
          format: uri
          example: "https://via.placeholder.com/150/cccccc/969696?text=User"
        phoneNumber:
          type: string
          example: "+1234567890"
        emailVerified:
          type: boolean
          example: true
        isActive:
          type: boolean
          example: true
        address:
          $ref: '#/components/schemas/Address'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: "2025-10-15T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-15T12:30:00.000Z"
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          default: "US"
          example: "US"
    UserPreferences:
      type: object
      properties:
        newsletter:
          type: boolean
          default: true
          example: true
        notifications:
          type: boolean
          default: true
          example: false
...